refactor(openomni): reconcile domain barrels with pattern a/b
