+++ packages/openomni/src/team/index.ts
@@ -1,6 +1,9 @@
 export { ApprovalGate } from "./approval-gate.js";
+export { EvaluationGate } from "./evaluation-gate.js";
 export { Teammate } from "./teammate.js";
 export { TeamOrchestrator } from "./team-orchestrator.js";
 export { ReviewLoop } from "./review-loop.js";
 export { StallDetector } from "./stall-detector.js";
 export { RunLedger } from "./run-ledger.js";
+export { TEAM_AGENTS, getAgentMetadata, resolveTeamAgent } from "./team-agents.js";
+export type { TeamAgentDefinition } from "./team-agents.js";
